+++ frontend/components/Home.js
@@ -1,7 +1,7 @@
 import React, {useState} from "react";
 import UploadFileInput from "./inputs/UploadFileInput";
 import SliderInstrument from "./instruments/SliderInstrument";
-import ShapeAnalysis from "./ShapeAnalysis";
+import ShapeAnalysis from "./textShapeToSound/ShapeAnalysis";
 
 const Home = () => {
     const [instrumentSamples, setInstrumentSamples] = useState(null);
