+++ frontend/components/textShapeToSound/ShapeAnalysis.js
